refactor(api): move ActivityRepository to school domain
