+++ api/tests/school/integration/domain/usecases/get-next-challenge_test.js
@@ -1,7 +1,7 @@
 import { databaseBuilder, expect, knex, mockLearningContent, sinon } from '../../../../test-helper.js';
 import { Assessment, Challenge } from '../../../../../lib/domain/models/index.js';
 import { Activity } from '../../../../../src/school/domain/models/Activity.js';
-import * as activityRepository from '../../../../../lib/infrastructure/repositories/activity-repository.js';
+import * as activityRepository from '../../../../../src/school/infrastructure/repositories/activity-repository.js';
 import * as assessmentRepository from '../../../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as challengeRepository from '../../../../../src/school/infrastructure/repositories/challenge-repository.js';
 import * as activityAnswerRepository from '../../../../../lib/infrastructure/repositories/activity-answer-repository.js';
